requirements.txt
genesis-world
torch>=2.4
ray[default]>=2.9
transformers>=4.52
accelerate>=1.7
diffusers>=0.34
sentencepiece
protobuf
huggingface_hub
mujoco
robot_descriptions
gradio>=5.0
anthropic
openai
moviepy>=2.0
imageio[ffmpeg]
numpy
Pillow
pyyaml
tqdm
